requirements.txt
Flask==2.3.2
Flask-Cors==3.1.1
torch==2.1.0
torchvision==0.15.0
Pillow==9.6.0
